requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2021.5.30=py36h06a4308_0
cudatoolkit=11.3.1=h2bc3f7f_2
dataclasses=0.8=pyh4f3eec9_6
ffmpeg=4.3=hf484d3e_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
intel-openmp=2022.1.0=h9e868ea_3769
jpeg=9e=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
pillow=8.3.1=py36h5aabda8_0
pip=21.2.2=py36h06a4308_0
python=3.6.13=h12debd9_1
pytorch=1.10.2=py3.6_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
readline=8.1.2=h7f8727e_1
setuptools=58.0.4=py36h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.2=h5082296_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.10.2=py36_cu113
torchvision=0.11.3=py36_cu113
typing_extensions=4.1.1=pyh06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h5eee18b_3
zstd=1.5.2=ha4553b6_0
